requirements.txt
google-genai
feedparser
python-dotenv
grpcio>=1.62.3,<1.68.0